requirements.txt
pyperclip==1.7.0
python-docx==0.8.10
selenium==3.141.0